UCT/TESTS: Wakeup tests fix for ugni
